requirements.txt
pygame>=2.1.0
Pillow>=8.0.0
requests>=2.25.0
opencv-python>=4.5.0
numpy>=1.21.0